fix(#305): fixing syntax errors and missing imports
