+++ ios/ble/central/CentralMangerDelegate.swift
@@ -1,0 +1,43 @@
+//
+//  CentralExtension.swift
+//  Openid4vpBle
+//
+//  Created by Shaik mohammed Jaffer on 03/01/23.
+//  Copyright © 2023 Facebook. All rights reserved.
+//
+
+import Foundation
+import CoreBluetooth
+import os
+
+extension Central: CBCentralManagerDelegate {
+    
+    func centralManagerDidUpdateState(_ central: CBCentralManager) {
+        switch central.state {
+        case .poweredOn:
+            print("Central Manager state is powered ON")
+            scanForPeripherals()
+        default:
+            print("Central Manager is in powered OFF")
+        }
+    }
+    
+    func centralManager(_ central: CBCentralManager, didDiscover peripheral: CBPeripheral, advertisementData: [String : Any], rssi RSSI: NSNumber) {
+        os_log("Discovered peripheral")
+        print("CBAdvertisementDataServiceUUIDsKey adv value " + String(describing: advertisementData[CBAdvertisementDataServiceUUIDsKey]))
+    }
+    
+    func centralManager(_ central: CBCentralManager, didConnect peripheral: CBPeripheral) {
+        os_log("Connected to peripheral: %@", String(describing: peripheral.name))
+        central.stopScan()
+    }
+    
+    func centralManager(_ central: CBCentralManager, didDisconnectPeripheral peripheral: CBPeripheral, error: Error?) {
+        os_log("Peripheral disconnected")
+        self.connectedPeripheral = nil
+    }
+    
+    func centralManager(_ central: CBCentralManager, didFailToConnect peripheral: CBPeripheral, error: Error?) {
+    }
+}
+
