+++ ios/ble/central/PeripheralDelegate.swift
@@ -1,0 +1,55 @@
+//
+//  PeripheralDelegate.swift
+//  Openid4vpBle
+//
+//  Created by Shaik mohammed Jaffer on 06/01/23.
+//  Copyright © 2023 Facebook. All rights reserved.
+//
+
+import Foundation
+import CoreBluetooth
+import os
+
+extension Central: CBPeripheralDelegate {
+    
+    func peripheral(_ peripheral: CBPeripheral, didDiscoverServices error: Error?) {
+        if let error = error {
+            os_log("Error while discovering services: %s", error.localizedDescription)
+            return
+        }
+        
+        guard let peripheralServices = peripheral.services else { return }
+        for service in peripheralServices {
+            peripheral.discoverCharacteristics([Peripheral.SERVICE_UUID, Peripheral.SCAN_RESPONSE_SERVICE_UUID], for: service)
+        }
+    }
+    
+    func peripheral(_ peripheral: CBPeripheral, didDiscoverCharacteristicsFor service: CBService, error: Error?) {
+        if let error = error {
+            os_log("Error discovering Characteristics: %s", error.localizedDescription)
+            return
+        }
+        
+        guard let serviceCharacteristics = service.characteristics else { return }
+        for characteristic in serviceCharacteristics {
+            peripheral.setNotifyValue(true, for: characteristic)
+        }
+    }
+    
+    func peripheral(_ peripheral: CBPeripheral, didUpdateValueFor characteristic: CBCharacteristic, error: Error?) {
+        if let error = error {
+            os_log("Unable to recieve updates from device: %s", error.localizedDescription)
+            return
+        }
+        
+        // use the new data from subscribed publisher
+    }
+    
+    func peripheral(_ peripheral: CBPeripheral, didWriteValueFor characteristic: CBCharacteristic, error: Error?) {
+        if let error = error {
+            os_log("Unable to write to characteristic: %@", error.localizedDescription)
+        }
+    }
+    
+    func peripheral(_ peripheral: CBPeripheral, didModifyServices invalidatedServices: [CBService]) {}
+}
